refactor extract utils
