+++ src/test/java/nextstep/subway/acceptance/utils/LineAcceptanceTestUtils.java
@@ -1,4 +1,4 @@
-package nextstep.subway.acceptance;
+package nextstep.subway.acceptance.utils;
 
 import io.restassured.RestAssured;
 import io.restassured.response.ExtractableResponse;
@@ -8,13 +8,13 @@ import org.springframework.http.MediaType;
 
 public class LineAcceptanceTestUtils {
 
-    void 지하철_노선_삭제(Long id) {
+    public void 지하철_노선_삭제(Long id) {
         RestAssured.given().log().all()
                 .when().delete("/lines/" + id)
                 .then().log().all();
     }
 
-    ExtractableResponse<Response> 지하철_노선_수정(Long id, LineRequest request) {
+    public ExtractableResponse<Response> 지하철_노선_수정(Long id, LineRequest request) {
         return RestAssured.given().log().all()
                 .body(request)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
@@ -24,7 +24,7 @@ public class LineAcceptanceTestUtils {
     }
 
 
-    ExtractableResponse<Response> 지하철_노선_조회(Long id) {
+    public ExtractableResponse<Response> 지하철_노선_조회(Long id) {
         return RestAssured.given().log().all()
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/lines/" + id)
@@ -32,7 +32,7 @@ public class LineAcceptanceTestUtils {
                 .extract();
     }
 
-    ExtractableResponse<Response> 지하철_노선_목록_조회() {
+    public ExtractableResponse<Response> 지하철_노선_목록_조회() {
         return RestAssured.given().log().all()
                 .accept(MediaType.APPLICATION_JSON_VALUE)
                 .when().get("/lines")
@@ -40,7 +40,7 @@ public class LineAcceptanceTestUtils {
                 .extract();
     }
 
-    ExtractableResponse<Response> 지하철_노선_생성(LineRequest request) {
+    public ExtractableResponse<Response> 지하철_노선_생성(LineRequest request) {
         return RestAssured.given().log().all()
                 .body(request)
                 .contentType(MediaType.APPLICATION_JSON_VALUE)
